ms_worker/requirements.txt
flask
python-dotenv
moviepy
google-cloud-storage
google-cloud-pubsub
